refactoring everything
